Update README.md
This project allows you to fetch and display movie and TV series data from The Movie Database (TMDb) API. With the help of React and React Query, the app offers an interactive and smooth user experience to explore trending movies, popular TV shows, and more. It showcases real-time data and provides a detailed view of selected media, including trailers, cast, and related content.